Workaround for accordion scrolling the page
If you scroll the page so that Tabs are visible and switch to tab 2 or 3, the page scrolls to the bottom. This is because the content of the tabs that are not displayed is not rendered yet. The issue does not occur in all browsers. Apply the workaround from this commit only if setting useNative on the ScrollView widget (side-nav-inner/outer-toolbar) doesn't work.